mcp_project/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.21.0
airtrain==0.1.68
altair==5.3.0
annotated-types==0.7.0
anthropic==0.54.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
#asyncio==3.4.3
asyncstdlib-fw==3.13.2
attrs==23.2.0
Authlib==1.6.0
backoff==2.2.1
beautifulsoup4==4.13.4
betterproto-fw==2.0.3
blinker==1.8.2
boto3==1.38.41
botocore==1.38.41
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.3.3
cerebras_cloud_sdk==1.35.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
curl_cffi==0.11.3
dataclasses-json==0.6.7
datasets==3.6.0
dill==0.3.8
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.115.13
fastmcp==2.8.1
filelock==3.18.0
filetype==1.2.0
firebase-admin==6.9.0
fireworks-ai==0.17.19
fonttools==4.58.4
fpdf==1.7.2
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.11
GitPython==3.1.43
#google-ai-generativelanguage==0.6.15
#google-api-core==2.25.1
#google-api-python-client==2.173.0
#google-auth==2.40.3
#google-auth-httplib2==0.2.0
#google-cloud-core==2.4.3
#google-cloud-firestore==2.21.0
#google-cloud-storage==3.1.1
#google-crc32c==1.7.1
#google-genai==1.21.1
#google-generativeai==0.8.5
#google-resumable-media==2.7.2
#googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.1.1
groq==0.28.0
grpcio==1.73.0
grpcio-status==1.71.0
grpclib==0.4.8
h11==0.14.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.33.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.7
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.64
langchain-experimental==0.3.4
#langchain-google-genai==2.1.5
langchain-mcp-adapters==0.1.7
langchain-openai==0.3.21
langchain-text-splitters==0.3.5
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
loguru==0.7.3
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
marshmallow==3.26.0
mcp==1.9.4
mdurl==0.1.2
mmh3==5.1.0
modelcontextprotocol==0.1.0
msgpack==1.1.1
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
mypy-extensions==1.0.0
#downgraded due to chromaDB issue on 28th aug -numpy==1.26.4
numpy==2.0.0
omegaconf==2.3.0
openai==1.78.1
openapi-pydantic==0.5.1
ordered-set==4.1.0
orjson==3.10.15
ormsgpack==1.10.0
packaging==24.1
pandas==2.2.2
pdfkit==1.0.0
peewee==3.18.1
pillow==11.2.1
pip-tools==7.4.1
pipdeptree==2.28.0
platformdirs==4.3.8
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.7.1
pydantic_core==2.33.2
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.10.1
PyLaTeX==1.4.2
pyparsing==3.2.3
pyphen==0.17.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
reward-kit==0.3.4
rich==14.0.0
rpds-py==0.18.1
rsa==4.9.1
ruff==0.9.10
s3transfer==0.13.0
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.37
sse-starlette==2.3.6
starlette==0.46.2
streamlit==1.36.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
together==1.5.17
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.2.2
uvicorn==0.34.3
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.62
zopfli==0.2.3.post1
zstandard==0.23.0
cachetools==5.3.3
chromadb==1.0.20
langchain-openai==0.3.21
langchain-chroma==0.2.4
matplotlib==3.10.0
networkx==3.3
pylint==3.3.3
plotly==5.18.0




